requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-multipart==0.0.12
typing-extensions>=4.12.2
openai==1.54.0
langchain==0.3.14
langchain-core==0.3.63
langchain-openai==0.3.0
